pyproject.toml
[tool.poetry]
name = "register-places"
version = "0.1.0"
description = ""
authors = ["shamada <shamada@jeol.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
googlemaps = "^4.6.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
mypy = "^0.931"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
